Made components modular
Only require components if they are configured
