+++ lua/slimline/utils.lua
@@ -1,0 +1,50 @@
+local M = {}
+
+--- Function to translate a mode into a string to show
+--- @return string
+function M.get_mode()
+  -- Note that: \19 = ^S and \22 = ^V.
+  local mode_map = {
+    ['n'] = 'NORMAL',
+    ['no'] = 'OP-PENDING',
+    ['nov'] = 'OP-PENDING',
+    ['noV'] = 'OP-PENDING',
+    ['no\22'] = 'OP-PENDING',
+    ['niI'] = 'NORMAL',
+    ['niR'] = 'NORMAL',
+    ['niV'] = 'NORMAL',
+    ['nt'] = 'NORMAL',
+    ['ntT'] = 'NORMAL',
+    ['v'] = 'VISUAL',
+    ['vs'] = 'VISUAL',
+    ['V'] = 'VISUAL',
+    ['Vs'] = 'VISUAL',
+    ['\22'] = 'VISUAL',
+    ['\22s'] = 'VISUAL',
+    ['s'] = 'SELECT',
+    ['S'] = 'SELECT',
+    ['\19'] = 'SELECT',
+    ['i'] = 'INSERT',
+    ['ic'] = 'INSERT',
+    ['ix'] = 'INSERT',
+    ['R'] = 'REPLACE',
+    ['Rc'] = 'REPLACE',
+    ['Rx'] = 'REPLACE',
+    ['Rv'] = 'VIRT REPLACE',
+    ['Rvc'] = 'VIRT REPLACE',
+    ['Rvx'] = 'VIRT REPLACE',
+    ['c'] = 'COMMAND',
+    ['cv'] = 'VIM EX',
+    ['ce'] = 'EX',
+    ['r'] = 'PROMPT',
+    ['rm'] = 'MORE',
+    ['r?'] = 'CONFIRM',
+    ['!'] = 'SHELL',
+    ['t'] = 'TERMINAL',
+  }
+
+  local mode = mode_map[vim.api.nvim_get_mode().mode] or 'UNKNOWN'
+  return mode
+end
+
+return M
